+++ drift_dev/lib/src/services/schema/schema_files.dart
@@ -1,3 +1,4 @@
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:collection/collection.dart';
@@ -13,10 +14,11 @@ import '../../analysis/resolver/shared/data_class.dart';
 import '../../analysis/results/results.dart';
 import '../../writer/utils/column_constraints.dart';
 import 'schema_isolate.dart';
+import 'sqlite_to_drift.dart';
 
 class _ExportedSchemaVersion {
-  static final Version current = _supportDialectSpecificConstraints;
-  static final Version _supportDialectSpecificConstraints = Version(1, 2, 0);
+  static final Version current = _exportSql;
+  static final Version _exportSql = Version(1, 3, 0);
   static final Version _supportDartIndex = Version(1, 1, 0);
 
   final Version version;
@@ -43,56 +45,30 @@ class SchemaWriter {
   /// Exports analyzed drift elements into a serialized format that can be used
   /// to re-construct the current database schema later.
   ///
-  /// Some drift elements, in particular Dart-defined views, are partially
-  /// defined at runtime and require running code. To infer the schema of these
-  /// elements, this method runs drift's code generator and spawns up a short-
-  /// lived isolate to collect the actual `CREATE` statements generated at
-  /// runtime.
+  /// Older versions of drift only exported schema information based on drift's
+  /// element model. This is sometimes inaccurate or incomplete, since generated
+  /// `CREATE` statements can depend on runtime code (e.g. for views where the
+  /// select statement is composed through drift's query builder).
+  ///
+  /// For this reason, we prefer to only export the `CREATE TABLE` statements
+  /// that drift actually generated as a reference. We still support the older
+  /// model, but the newer is much simpler while also being more reliable.
   Future<Map<String, Object?>> createSchemaJson({File? dumpStartupCode}) async {
-    final requiresRuntimeInformation = <DriftSchemaElement>[];
-    for (final element in elements) {
-      switch (element) {
-        case DriftTable():
-          for (final column in element.columns) {
-            if (column.sqlType is ColumnCustomType) {
-              requiresRuntimeInformation.add(element);
-              continue;
-            }
-
-            if (column.defaultArgument != null) {
-              // This is an arbitrary Dart expression allowed to contain user
-              // code. To make sure the schema file stays valid, evaluate it
-              // once now and replace the expression with the result as a
-              // constant when serializing.
-              requiresRuntimeInformation.add(element);
-              continue;
-            }
-          }
-        case DriftView():
-          if (element.source is! SqlViewSource) {
-            requiresRuntimeInformation.add(element);
-          }
-      }
-    }
-
     final knownStatements = <String, List<(SqlDialect, String)>>{};
-    if (requiresRuntimeInformation.isNotEmpty) {
-      try {
-        final statements = await SchemaIsolate.collectStatements(
-          options: options,
-          allElements: elements,
-          elementFilter: requiresRuntimeInformation,
-          dumpStartupCode: dumpStartupCode,
-        );
+    try {
+      final statements = await SchemaIsolate.collectStatements(
+        options: options,
+        allElements: elements,
+        dumpStartupCode: dumpStartupCode,
+      );
 
-        for (final statement in statements) {
-          knownStatements
-              .putIfAbsent(statement.elementName, () => [])
-              .add((statement.dialect, statement.createStatement));
-        }
-      } on SchemaIsolateException catch (e) {
-        _logger.warning(e.description(isFatal: false));
+      for (final statement in statements) {
+        knownStatements
+            .putIfAbsent(statement.elementName, () => [])
+            .add((statement.dialect, statement.createStatement));
       }
+    } on SchemaIsolateException catch (e) {
+      _logger.warning(e.description(isFatal: false));
     }
 
     return {
@@ -103,9 +79,24 @@ class SchemaWriter {
       },
       'options': _serializeOptions(),
       'entities': elements
-          .map((e) => _entityToJson(e, knownStatements))
+          .map((e) => _entityToJson(e,
+              e is DriftSchemaElement ? knownStatements[e.schemaName] : null))
           .whereType<Map>()
           .toList(),
+      if (knownStatements.isNotEmpty)
+        'fixed_sql': [
+          for (final MapEntry(:key, :value) in knownStatements.entries)
+            {
+              'name': key,
+              'sql': [
+                for (final (dialect, sql) in value)
+                  {
+                    'dialect': dialect.name,
+                    'sql': sql,
+                  }
+              ]
+            }
+        ],
     };
   }
 
@@ -117,8 +108,8 @@ class SchemaWriter {
     return asJson;
   }
 
-  Map<String, Object?>? _entityToJson(DriftElement entity,
-      Map<String, List<(SqlDialect, String)>> knownStatements) {
+  Map<String, Object?>? _entityToJson(
+      DriftElement entity, List<(SqlDialect, String)>? knownSql) {
     String? type;
     Map<String, Object?>? data;
 
@@ -132,8 +123,8 @@ class SchemaWriter {
       // a different thing when dependencies are changed, while we want an
       // immutable schema snapshot.
       CreateTableStatement? actualTable;
-      if (knownStatements[entity.schemaName] case final known?) {
-        final sql = known.firstWhere((e) => e.$1 == SqlDialect.sqlite).$2;
+      if (knownSql != null) {
+        final sql = knownSql.firstWhere((e) => e.$1 == SqlDialect.sqlite).$2;
         final engine = SqlEngine(EngineOptions(version: SqliteVersion.current));
 
         final result = engine.parse(sql);
@@ -171,8 +162,8 @@ class SchemaWriter {
       };
     } else if (entity is DriftView) {
       String? sql;
-      if (knownStatements[entity.schemaName] case final known?) {
-        sql = known.firstWhere((e) => e.$1 == SqlDialect.sqlite).$2;
+      if (knownSql != null) {
+        sql = knownSql.firstWhere((e) => e.$1 == SqlDialect.sqlite).$2;
       } else {
         final source = entity.source;
         if (source is! SqlViewSource) {
@@ -340,6 +331,9 @@ class SchemaWriter {
   }
 
   static final _logger = Logger('drift_dev.SchemaWriter');
+
+  /// A suitable JSON encoder for emitting schema export JSONs.
+  static const json = JsonEncoder.withIndent('  ');
 }
 
 /// Reads files generated by [SchemaWriter].
@@ -359,13 +353,15 @@ class SchemaReader {
 
   SchemaReader._();
 
-  factory SchemaReader.readJson(Map<String, dynamic> json) {
-    return SchemaReader._().._read(json);
+  static Future<SchemaReader> readJson(Map<String, dynamic> json) async {
+    final reader = SchemaReader._();
+    await reader._read(json);
+    return reader;
   }
 
   Iterable<DriftElement> get entities => _entitiesById.values;
 
-  void _read(Map<String, dynamic> json) {
+  Future<void> _read(Map<String, dynamic> json) async {
     final meta = json['_meta'] as Map<String, Object?>;
     _version = _ExportedSchemaVersion(Version.parse(meta['version'] as String));
 
@@ -375,8 +371,14 @@ class SchemaReader {
         {
           'store_date_time_values_as_text': false,
         };
-
     final entities = json['entities'] as List<dynamic>;
+
+    if (json['fixed_sql'] case List<Object?> fixedSql) {
+      // If we have access to actual CREATE statements drift used to generate at
+      // the version the schema was created, that is much better than
+      // reconstructing from element exports. Do that.
+      return _processFromSql(fixedSql, entities);
+    }
 
     for (final raw in entities) {
       final rawData = raw as Map<String, dynamic>;
@@ -396,6 +398,48 @@ class SchemaReader {
 
   DriftDeclaration get _declaration =>
       DriftDeclaration(elementUri, -1, '<unknown>');
+
+  Future<void> _processFromSql(
+      List<Object?> fixedSql, List<Object?> entities) async {
+    final sql = <String>[];
+
+    for (final entry in fixedSql) {
+      for (final dialect in (entry as Map)['sql'] as List<Object?>) {
+        if ((dialect as Map)['dialect'] == SqlDialect.sqlite.name) {
+          sql.add(dialect['sql'] as String);
+        }
+      }
+    }
+
+    final elements = await extractDriftElementsFromSql(sql);
+    for (final (i, element) in elements.indexed) {
+      if (element is DriftTable) {
+        final pascalCase = ReCase(element.id.name).pascalCase;
+        // Mirror the naming scheme from _readTable when reading from SQL.
+        element
+          ..baseDartName = pascalCase
+          ..fixedEntityInfoName = pascalCase
+          ..nameOfRowClass = '${pascalCase}Data';
+      } else if (element is DriftView) {
+        for (final entity in entities.cast<Map<String, Object?>>()) {
+          if (entity['type'] == 'view') {
+            final data = entity['data'] as Map<String, Object?>;
+            if (data['name'] == element.schemaName) {
+              final entityInfoName = data['dart_info_name'] as String;
+
+              element
+                ..entityInfoName = data['dart_info_name'] as String
+                ..nameOfRowClass = dataClassNameForClassName(entityInfoName);
+
+              break;
+            }
+          }
+        }
+      }
+
+      _entitiesById[i] = element;
+    }
+  }
 
   void _processById(int id) {
     if (_entitiesById.containsKey(id)) return;